requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

Scrapy == 1.5.0

Twisted == 17.5.0

demjson == 2.2.4

elasticsearch-dsl == 6.1.0

kafka_python == 1.4.3

openpyxl == 2.4.7

selenium == 3.5.0

xlrd == 1.1.0

pandas == 0.23.3

pytdx == 1.55

apscheduler == 3.4.0

requests == 2.18.4

ccxt == 1.14.224

Flask == 1.0

marshmallow

schedule

pymongo